Remodel the interface with window managers
After looking at implementing the window surface creation, it seems that
keeping all Vulkan code within the renderer isn't necessarily optimal.
It would require the caller to pass platform-dependant window handles
(e.g.: `xcb_connection_t *` and `xcb_window_t` for Linux, `HINSTANCE`
and `HWND` for Windows, ...), which is both inelegant and requires
defining new members to the struct `RendererCreateInfo` with types that
can only be retrieved by including the humongous headers of the window
manager they belong to. And we really want to keep header dependencies
to a minimum in our own headers.

In such scenarios where two different modules (renderer and window
manager here) depend on each other, it seems that the best way to go
while keeping coupling explicit and to a minimum is to use callbacks.

Hence the Vulkan operations that are related to a window manager
(e.g.: instance extension names, window surface creation, ...) are to be
delegated to the caller which is given the responsability to provide
callbacks interfacing the renderer's back-end with the window manager.

It seems that GLFW and SDL already assume and facilitate this approach.