Support the creation of a standalone renderer
This is done by passing `NULL` for the window manager interface.

If a valid object is passed for the window manager interface, the required
instance extensions for interfacing with the surface, such as
'VK_KHR_surface', are enabled. Another function not yet implemented
will allow to bind the renderer with the window manager once both
objects are created.