Implement the surface creation
When creating a logical device, there are two ways of checking if some
queue families from a physical device support presenting images to a
surface:

- `vkGetPhysicalDeviceSurfaceSupportKHR()`, which requires a surface to
  be already created.
- `vkGetPhysicalDevice*PresentationSupportKHR()`, which usually "only"
  requires some platform-specific handles or nothing at all in the case
  of Windows.

Ideally the latter option is best since it allows delaying the creation
of the surface (window) to when both the renderer and the window manager
objects are created, as initially planned in commit
8559a5f50de0968a7a5a037e4ce524c728058562.

Since querying presentation support falls into the window manager side
of things, it is best addressed with callbacks to be provided by the
user. Unfortunately, only GLFW provides a function to easily write such
a callback while much more work would be required to create one with
SDL2 (including writing platform-specific code).

Hence, let's stick here with the common denominator for now and give up
on the idea of deferred window creation. If the renderer is to present
images to a surface, then its window needs to be already created.