Fix fbs having a different extent than the images
Upon resizing the surface, the swap chain was created using the current
capabilities extent derived from a recent call to
`vkGetPhysicalDeviceSurfaceCapabilitiesKHR()` while, on the other hand,
the frame buffers were initialized with the extent coming from the
new size of the surface.

In some occasions, both extents turned out to have different values,
hence a distinction is now being made between 'surface extent', that is
the desired value of the image extent based on the surface size, and
'image extent', that is the actual extent available for the swap chain
images.

Only the image extent is now used to initialize all Vulkan objects.