Switch to C99
Restricting to C89 was done primarily for educational reasons but, as
the project goes on, the lack of some features is regrettable, if only
to be in line with the use of the `DK_USE_STD_*_TYPES` macros, to have
access to the `[u]intptr_t` type for bit mask operations on pointers and
`long long` for larger (mostly 64-bit) integer types, and to simplify
the logging macros with variadic arguments.

Switching to C11 is also considered for its memory alignment features
and for its multithreading support.

Still, outside the features previously mentioned, the plan is to keep
most of the code compatible with C89.