Move the callbacks into the `*CreateInfo` structs
It would make sense to keep them as separate parameters if different
functions affecting a same object (e.g. `DkRenderer`) would each allow
to take a different set of callbacks, which is not the case here for now
since the same allocator and logger passed at creation time is used
through the whole object's life.