Add aligned allocation callbacks
Firstly, the non-aligned allocation callbacks are kept instead of
replacing them with their aligned counterpart because implementations
for the alignment version most likely come with an overhead and it
should be up to the user to decide if this overhead is acceptable or
not.

Secondly, the default implementation of the aligned allocation callbacks
does not use use C11's `aligned_alloc()` because the standard doesn't
provide an aligned version of `realloc()` and implementing a custom
reallocator on top of memory blocks allocated with `aligned_alloc()`
doesn't seem optimal since copying would always be required instead of
trying to expand blocks in place. Not providing any reallocation
callback was considered but having one could come in handy and Vulkan
does require one to be implemented if we want to override their own
allocation callbacks for logging purposes or other reasons.