Revamp the Makefile
Upon wanting to restructure the project and to add more executables
to it, it was clear that a more scalable Makefile was required.

Not only new playgrounds are now automatically taken into account but
it is also possible to build for multiple combinations of
configurations and architectures at once.

That is, running `make config="debug release" arch="x86-64 corei7"`
generates 4 different builds: debug x86-64, debug corei7,
release x86-64, and release corei7.

On top of that, a C++ compilation can be chosen in place of the C one
and the output directory for the build can also be changed.