Revert the output value initialization
This was initially done to prevent a segfault but the culprit was
in fact the calling code which has been fixed in the commit
a0777726c4d71fa4dd15e5b9d945c9f9c5d4a9ef.

Returning an error result from `Create*()` functions should be enough
to signal that the object hasn't been initialized and shouldn't be
passed to `Destroy*()`.