Promote two variables to parameters
This is to provide more flexibility when the source files aren't
meant to be used as the `CREATE_OBJECT_RULES` and `CREATE_BINARY_RULES`
functions hardcoded it.

For example, all the files found in a demo folder might compile to a
single binary executable, whereas a test folder might define multiple
executables, thus requiring different source files and target name
from the rest of the project.