Remove typedefs
Typedefing structs is troublesome whenever forward declarations are
needed since typedef redefinitions are not allowed when compiling in
strict ANSI mode.

Also the strategy of grouping all the typedefs in a single place was
bad design since it removed the need for explicit forward declarations
altogether by simply including a single file.